adicionando edição de caso e ongs

+++ frontend/src/pages/Profile/index.js
@@ -1,10 +1,10 @@
-import React, {useEffect, useState} from 'react'
-
+import React, { useEffect, useState } from 'react'
+import { FiEdit, FiPower, FiTrash2 } from 'react-icons/fi'
+import { Link, useHistory } from 'react-router-dom'
 import logo from '../../assets/logo.svg'
-import {FiPower, FiTrash2} from 'react-icons/fi'
-import {Link, useHistory} from 'react-router-dom'
 import api from '../../services/api'
 import './styles.css'
+
 
 export default function Profile() {
   const ongName = localStorage.getItem('ongName')
@@ -14,6 +14,11 @@ export default function Profile() {
   const history = useHistory()
 
   useEffect(() => {
+    if (!ongId) {
+      history.push('/')
+      return;
+    }
+
     api.get('profile', {
       headers: {
         Authorization: ongId
@@ -21,20 +26,25 @@ export default function Profile() {
     }).then(resp => {
       setData(resp.data)
     })
-  }, [ongId])
+  }, [ongId, history])
 
-  async function handleDelete(id) {
-    try {
-      await api.delete(`incidents/${id}`, {
+  function handleDelete(id) {
+    if (window.confirm('Deseja realmente excluir este caso?')) {
+      api.delete(`incidents/${id}`, {
         headers: {
           Authorization: ongId
         }
+      }).then(resp => {
+        setData(data.filter(incident => incident.id !== id))
+      }).catch(error => {
+        alert('Não foi possível excluir o caso')
       })
-
-      setData(data.filter(incident => incident.id !== id))
-    } catch (error) {
-      alert('Não foi possível excluir o caso')
     }
+  }
+
+  function handleEdit(incident) {
+    const {id, title, description, value} = incident
+    history.push(`/incidents/update`, {id, title, description, value})
   }
 
   function handleLogout() {
@@ -46,7 +56,9 @@ export default function Profile() {
     <div className="profile-container">
       <header>
         <img src={logo} alt="Be The Hero"/>
-        <span>Bem vinda, {ongName}</span>
+        <span>
+          Bem vinda, <Link to="/register">{ongName}</Link>
+        </span>
 
         <Link to="/incidents/new" className="button">
           Cadastrar novo caso
@@ -74,6 +86,10 @@ export default function Profile() {
 
               <button onClick={() => handleDelete(incident.id)} type="button">
                 <FiTrash2 color="#a8a8b3" size="20" />
+              </button>
+
+              <button onClick={() => handleEdit(incident)} type="button">
+                <FiEdit color="#a8a8b3" size="20" />
               </button>
             </li>
           ))}
